- Rename model : mobilenet -> mobilenet_v1 - Update README.md in x86 - Add model_download.sh to download models from each model folder - Separate Dockerfile for each API
