+++ bench/mobilenet_v1/rest_bench.py
@@ -1,5 +1,5 @@
 #preprocessing library
-from mobilenet import preprocessing
+from mobilenet_v1 import preprocessing
 import numpy as np
 
 #REST 요청 관련 library
@@ -10,7 +10,6 @@ import json
 import concurrent.futures
 
 def run_bench(num_tasks, server_address):
-    # 저장할 google spread sheet id
     model_name = "mobilenet_v1"
     image_file_path = "../../../dataset/imagenet/imagenet_1000_raw/n01843383_1.JPEG"
