+++ x86/README.md
@@ -2,5 +2,18 @@
 
 - 각 모델 Docker container 빌드 전, build하는 호스트에서 model_download.sh를 참고하여 model을 다운로드 후 진행해야 합니다.
 
-- Docker Image build시, Download한 model 폴더를 같은 Directory에 복사 후 진행해야 합니다.
-  - Edge 장비의 경우, 각각 model을 COPY 후 빌드할 필요 없이, docker -v option을 사용하여도 됩니다.
+- Docker Image build 전, 같은 폴더에 위치한 model_download.sh를 실행 후 진행해야 합니다.
+  ```shell
+  chmod +x ./model_download.sh && ./model_download.sh
+  ```
+
+- Docker bulld & run 방법 (예시)
+  ```shell
+  tag="tfserving-cloud"
+  version="latest"
+  API="gRPC" #gPRC or REST
+  HOST_PORT=8080
+  CONTAINER_PORT=8500 #gRPC = 8500, REST = 8501
+  docker build -t $tag:$version -f Dockerfile-$API .
+  docker run -it -p $HOST_PORT:$CONTAINER_PORT $tag:$version
+  ```
